requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==35.0.0
geographiclib==1.52
geopy==2.2.0
h11==0.12.0
idna==3.3
importchecker==2.0
mysql-client==0.0.1
mysql-connector-python==8.0.27
outcome==1.1.0
protobuf==3.18.1
pycparser==2.20
pyOpenSSL==21.0.0
python-slugify==5.0.2
requests==2.27.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.2
wsproto==1.0.0
